[REFACTOR] Extract SymbolicClient base class from sanitizer and race detector
SymbolicSanitizer and RaceDetector shared ~500 lines of duplicated code:
operation overrider functions, numpy-to-symbolic-op mappings, RangeWrapper,
and for-loop callback infrastructure. This extracts the shared code into a
SymbolicClient(Client) base class in symbolic_engine.py, with subclass hooks
(_on_data_dependent_value, _should_skip_loop_hooks, _loop_hook_before/after)
for domain-specific behavior.
