+++ triton_viz/clients/__init__.py
@@ -2,7 +2,7 @@ from .profiler.profiler import Profiler
 from .profiler.data import LoadStoreBytes, OpTypeCounts
 from .sanitizer.sanitizer import Sanitizer
 from .sanitizer.data import OutOfBoundsRecord
-from .symbolic_engine import SymbolicExpr
+from .symbolic_engine import SymbolicExpr, SymbolicClient, RangeWrapper
 from .tracer.tracer import Tracer
 from .race_detector.race_detector import RaceDetector
 from .race_detector.data import RaceType, RaceRecord
@@ -14,6 +14,8 @@ __all__ = [
     "OpTypeCounts",
     "OutOfBoundsRecord",
     "SymbolicExpr",
+    "SymbolicClient",
+    "RangeWrapper",
     "Tracer",
     "RaceDetector",
     "RaceType",
