+++ triton_viz/clients/race_detector/race_detector.py
@@ -1,114 +1,33 @@
-from collections.abc import Callable, Iterator
-from dataclasses import dataclass
+from collections.abc import Callable
 from typing import Any, Optional
 from collections import defaultdict
 
 import numpy as np
 from z3 import Solver, Int, And, Or, sat, substitute
-from z3.z3 import ArithRef, IntNumRef
+from z3.z3 import ArithRef
 
-import triton.language as tl
-
-from ...core.client import Client
-from ...core.callbacks import OpCallbacks, ForLoopCallbacks
+from ...core.callbacks import OpCallbacks
 from ...core.data import (
     Op,
     RawLoad,
     Load,
     RawStore,
     Store,
-    UnaryOp,
-    BinaryOp,
-    TernaryOp,
-    ProgramId,
-    Dot,
-    MakeRange,
-    AddPtr,
-    ExpandDims,
-    Broadcast,
-    ReduceSum,
-    ReduceMax,
-    ReduceMin,
-    Splat,
     MakeBlockPointer,
     TensorPointerLoad,
     TensorPointerStore,
-    Idiv,
-    Rsqrt,
-    CastImpl,
-    Reshape,
-    Trans,
-    Join,
-    Fabs,
-    Ashr,
-    Advance,
-    FpToFp,
-    Umulhi,
-    CumSum,
-    Bitcast,
     AtomicCas,
     AtomicRMW,
 )
 from ..symbolic_engine import (
     SymbolicExpr,
-    SymbolicExprDataWrapper,
-    LoopContext,
+    SymbolicClient,
 )
 from .data import AccessType, MemoryAccess, SymbolicMemoryAccess, RaceRecord, RaceType
 from ...utils.traceback_utils import extract_user_frames
 
 
-_UNARY_NUMPY_TO_SYM_OP: dict[Callable[..., Any], str] = {
-    np.cos: "cos",
-    np.exp: "exp",
-    np.exp2: "exp2",
-    np.abs: "abs",
-    np.floor: "floor",
-    np.ceil: "ceil",
-    np.log: "log",
-    np.log2: "log2",
-    np.sqrt: "sqrt",
-    np.sin: "sin",
-}
-
-_BINARY_NUMPY_TO_SYM_OP: dict[Callable[..., Any], str] = {
-    np.add: "add",
-    np.subtract: "sub",
-    np.multiply: "mul",
-    np.divide: "div",
-    np.less: "less",
-    np.less_equal: "less_equal",
-    np.greater: "greater",
-    np.greater_equal: "greater_equal",
-    np.not_equal: "not_equal",
-    np.equal: "equal",
-    np.fmod: "mod",
-    np.maximum: "maximum",
-    np.minimum: "minimum",
-    np.bitwise_and: "bitwise_and",
-    np.bitwise_or: "bitwise_or",
-    np.bitwise_xor: "bitwise_xor",
-    np.right_shift: "right_shift",
-    np.left_shift: "left_shift",
-}
-
-
-@dataclass
-class _RangeWrapper:
-    iterable: Any
-    length: int
-    start: int
-    stop: int
-    step: int
-
-    def __iter__(self) -> Iterator[Any]:
-        return iter(self.iterable)
-
-    def __len__(self) -> int:
-        return self.length
-
-
-class RaceDetector(Client):
+class RaceDetector(SymbolicClient):
     NAME = "race_detector"
 
     def __init__(self):
@@ -121,10 +40,6 @@ class RaceDetector(Client):
         # Each entry: (OpCallbacks, original_overrider_fn)
         self._op_callbacks: list[tuple[OpCallbacks, Callable]] = []
         self._races: list[RaceRecord] = []
-        self._loop_stack: list[LoopContext] = []
-        SymbolicExpr.set_loop_ctx_provider(
-            lambda *_a, **_kw: self._loop_stack[-1] if self._loop_stack else None
-        )
 
     # ── Client interface ─────────────────────────────────────────
 
@@ -171,228 +86,108 @@ class RaceDetector(Client):
     def grid_idx_callback(self, grid_idx: tuple[int, ...]):
         self.grid_idx = grid_idx
 
+    # ── Race-detector-specific operation overriders ─────────────
+
+    def _op_raw_load_overrider(self, ptr, cache_modifier, eviction_policy, is_volatile):
+        return self._op_load_overrider(
+            ptr, None, None, cache_modifier, eviction_policy, is_volatile
+        )
+
+    def _op_load_overrider(self, ptr, mask, other, *args):
+        if isinstance(ptr, SymbolicExpr) and ptr.has_op("load"):
+            self._need_concrete_fallback = True
+            ptr = ptr.replace_subtree("load")
+        if isinstance(mask, SymbolicExpr) and mask.has_op("load"):
+            self._need_concrete_fallback = True
+            mask = mask.replace_subtree("load")
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        mask_sym = SymbolicExpr.from_value(mask) if mask is not None else None
+        other_sym = SymbolicExpr.from_value(other) if other is not None else None
+        ret = SymbolicExpr.create("load", ptr_sym, mask_sym, other_sym)
+        self._record_symbolic_access(AccessType.LOAD, ptr_sym, mask_sym)
+        return ret
+
+    def _op_raw_store_overrider(self, ptr, value, cache_modifier, eviction_policy):
+        return self._op_store_overrider(
+            ptr, value, None, cache_modifier, eviction_policy
+        )
+
+    def _op_store_overrider(self, ptr, value, mask, *args):
+        if isinstance(ptr, SymbolicExpr) and ptr.has_op("load"):
+            self._need_concrete_fallback = True
+            ptr = ptr.replace_subtree("load")
+        if isinstance(mask, SymbolicExpr) and mask.has_op("load"):
+            self._need_concrete_fallback = True
+            mask = mask.replace_subtree("load")
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        value_sym = SymbolicExpr.from_value(value)
+        mask_sym = SymbolicExpr.from_value(mask) if mask is not None else None
+        ret = SymbolicExpr.create("store", ptr_sym, value_sym, mask_sym)
+        self._record_symbolic_access(AccessType.STORE, ptr_sym, mask_sym)
+        return ret
+
+    def _op_make_block_ptr_overrider(
+        self, base, shape, strides, offsets, tensor_shape, order
+    ):
+        raise NotImplementedError("MakeBlockPtr is not supported yet.")
+
+    def _op_tensor_pointer_load_overrider(
+        self,
+        ptr,
+        boundary_check,
+        padding_option,
+        cache_modifier,
+        eviction_policy,
+        is_volatile,
+    ):
+        raise NotImplementedError("TensorPointerLoad is not supported yet.")
+
+    def _op_tensor_pointer_store_overrider(
+        self, ptr, value, boundary_check, cache_modifier, eviction_policy
+    ):
+        raise NotImplementedError("TensorPointerStore is not supported yet.")
+
+    def _op_atomic_cas_overrider(self, ptr, cmp, val, sem, scope):
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        cmp_sym = SymbolicExpr.from_value(cmp)
+        val_sym = SymbolicExpr.from_value(val)
+        ret = SymbolicExpr.create("atomic_cas", ptr_sym, cmp_sym, val_sym)
+        self._record_symbolic_access(AccessType.ATOMIC, ptr_sym, None)
+        return ret
+
+    def _op_atomic_rmw_overrider(self, rmwOp, ptr, val, mask, sem, scope):
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        val_sym = SymbolicExpr.from_value(val)
+        mask_sym = SymbolicExpr.from_value(mask)
+        ret = SymbolicExpr.create("atomic_rmw", ptr_sym, val_sym, mask_sym)
+        self._record_symbolic_access(AccessType.ATOMIC, ptr_sym, mask_sym)
+        return ret
+
+    # ── For-loop hook overrides ──────────────────────────────────
+
+    def _on_data_dependent_value(self) -> None:
+        self._need_concrete_fallback = True
+
+    def _should_skip_loop_hooks(self) -> bool:
+        return self._phase != "symbolic"
+
     def register_op_callback(self, op_type: type[Op]) -> OpCallbacks:
-        # ── Symbolic overriders (active during symbolic phase) ────
+        overrider_map = self._build_op_overrider_map()
+        overrider_map.update(
+            {
+                RawLoad: self._op_raw_load_overrider,
+                Load: self._op_load_overrider,
+                RawStore: self._op_raw_store_overrider,
+                Store: self._op_store_overrider,
+                MakeBlockPointer: self._op_make_block_ptr_overrider,
+                TensorPointerLoad: self._op_tensor_pointer_load_overrider,
+                TensorPointerStore: self._op_tensor_pointer_store_overrider,
+                AtomicCas: self._op_atomic_cas_overrider,
+                AtomicRMW: self._op_atomic_rmw_overrider,
+            }
+        )
 
-        def op_program_id_overrider(axis):
-            return SymbolicExpr.create("pid", axis)
-
-        def op_raw_load_overrider(ptr, cache_modifier, eviction_policy, is_volatile):
-            return op_load_overrider(
-                ptr, None, None, cache_modifier, eviction_policy, is_volatile
-            )
-
-        def op_load_overrider(ptr, mask, other, *args):
-            if isinstance(ptr, SymbolicExpr) and ptr.has_op("load"):
-                self._need_concrete_fallback = True
-                ptr = ptr.replace_subtree("load")
-            if isinstance(mask, SymbolicExpr) and mask.has_op("load"):
-                self._need_concrete_fallback = True
-                mask = mask.replace_subtree("load")
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            mask_sym = SymbolicExpr.from_value(mask) if mask is not None else None
-            other_sym = SymbolicExpr.from_value(other) if other is not None else None
-            ret = SymbolicExpr.create("load", ptr_sym, mask_sym, other_sym)
-            self._record_symbolic_access(AccessType.LOAD, ptr_sym, mask_sym)
-            return ret
-
-        def op_raw_store_overrider(ptr, value, cache_modifier, eviction_policy):
-            return op_store_overrider(ptr, value, None, cache_modifier, eviction_policy)
-
-        def op_store_overrider(ptr, value, mask, *args):
-            if isinstance(ptr, SymbolicExpr) and ptr.has_op("load"):
-                self._need_concrete_fallback = True
-                ptr = ptr.replace_subtree("load")
-            if isinstance(mask, SymbolicExpr) and mask.has_op("load"):
-                self._need_concrete_fallback = True
-                mask = mask.replace_subtree("load")
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            value_sym = SymbolicExpr.from_value(value)
-            mask_sym = SymbolicExpr.from_value(mask) if mask is not None else None
-            ret = SymbolicExpr.create("store", ptr_sym, value_sym, mask_sym)
-            self._record_symbolic_access(AccessType.STORE, ptr_sym, mask_sym)
-            return ret
-
-        def op_atomic_rmw_overrider(rmwOp, ptr, val, mask, sem, scope):
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            val_sym = SymbolicExpr.from_value(val)
-            mask_sym = SymbolicExpr.from_value(mask)
-            ret = SymbolicExpr.create("atomic_rmw", ptr_sym, val_sym, mask_sym)
-            # Record ptr_expr (not the full atomic node) to avoid Z3 NotImplementedError
-            self._record_symbolic_access(AccessType.ATOMIC, ptr_sym, mask_sym)
-            return ret
-
-        def op_atomic_cas_overrider(ptr, cmp, val, sem, scope):
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            cmp_sym = SymbolicExpr.from_value(cmp)
-            val_sym = SymbolicExpr.from_value(val)
-            ret = SymbolicExpr.create("atomic_cas", ptr_sym, cmp_sym, val_sym)
-            self._record_symbolic_access(AccessType.ATOMIC, ptr_sym, None)
-            return ret
-
-        def op_unary_op_overrider(arg, op):
-            arg_sym = SymbolicExpr.from_value(arg)
-            name = _UNARY_NUMPY_TO_SYM_OP.get(op)
-            if name is None:
-                raise NotImplementedError(f"Unsupported unary operation: {op}")
-            return SymbolicExpr.create(name, arg_sym)
-
-        def op_binary_op_overrider(lhs, rhs, op):
-            lhs_sym = SymbolicExpr.from_value(lhs)
-            rhs_sym = SymbolicExpr.from_value(rhs)
-            op_name = _BINARY_NUMPY_TO_SYM_OP.get(op)
-            if op_name is None:
-                raise NotImplementedError(f"Unsupported binary operation: {op}")
-            return SymbolicExpr.create(op_name, lhs_sym, rhs_sym)
-
-        def op_ternary_op_overrider(lhs, rhs, other, op):
-            lhs_sym = SymbolicExpr.from_value(lhs)
-            rhs_sym = SymbolicExpr.from_value(rhs)
-            other_sym = SymbolicExpr.from_value(other)
-            if op is np.where:
-                return SymbolicExpr.create("where", lhs_sym, rhs_sym, other_sym)
-            raise NotImplementedError(f"Unsupported ternary operation: {op}")
-
-        def op_addptr_overrider(ptr, offset):
-            return SymbolicExpr.create(
-                "addptr",
-                SymbolicExpr.from_value(ptr),
-                SymbolicExpr.from_value(offset),
-            )
-
-        def op_dot_overrider(a, b, d, input_precision, max_num_imprecise_acc):
-            a_sym = SymbolicExpr.from_value(a)
-            b_sym = SymbolicExpr.from_value(b)
-            d_sym = SymbolicExpr.from_value(d) if d is not None else None
-            return SymbolicExpr.create("dot", a_sym, b_sym, d_sym)
-
-        def op_make_range_overrider(ret_ty, start, end):
-            return SymbolicExpr.create(
-                "arange",
-                SymbolicExpr.from_value(ret_ty),
-                SymbolicExpr.from_value(start),
-                SymbolicExpr.from_value(end),
-            )
-
-        def op_expand_dims_overrider(arg, axis):
-            return SymbolicExpr.create(
-                "expand_dims", SymbolicExpr.from_value(arg), axis
-            )
-
-        def op_broadcast_overrider(arg, shape):
-            return SymbolicExpr.create("broadcast", SymbolicExpr.from_value(arg), shape)
-
-        def op_reduce_sum_overrider(input, axis=None, keep_dims=False, **kwargs):
-            return SymbolicExpr.create(
-                "sum", SymbolicExpr.from_value(input), axis, keep_dims
-            )
-
-        def op_reduce_max_overrider(input, axis=None, keep_dims=False, **kwargs):
-            return SymbolicExpr.create(
-                "max", SymbolicExpr.from_value(input), axis, keep_dims
-            )
-
-        def op_reduce_min_overrider(input, axis=None, keep_dims=False, **kwargs):
-            return SymbolicExpr.create(
-                "min", SymbolicExpr.from_value(input), axis, keep_dims
-            )
-
-        def op_splat_overrider(shape, arg):
-            return SymbolicExpr.create("splat", shape, SymbolicExpr.from_value(arg))
-
-        def op_idiv_overrider(lhs, rhs):
-            return SymbolicExpr.from_value(lhs) // SymbolicExpr.from_value(rhs)
-
-        def op_rsqrt_overrider(arg):
-            return SymbolicExpr.create("rsqrt", SymbolicExpr.from_value(arg))
-
-        def op_cast_impl_overrider(src, dst_type):
-            return SymbolicExpr.create("cast_impl", src, dst_type)
-
-        def op_reshape_overrider(arg, shape, allow_reorder):
-            return SymbolicExpr.create(
-                "reshape",
-                SymbolicExpr.from_value(arg),
-                SymbolicExpr.from_value(shape),
-            )
-
-        def op_trans_overrider(arg, perm=[1, 0]):
-            return SymbolicExpr.create("trans", SymbolicExpr.from_value(arg), perm)
-
-        def op_join_overrider(lhs, rhs):
-            return SymbolicExpr.create(
-                "join",
-                SymbolicExpr.from_value(lhs),
-                SymbolicExpr.from_value(rhs),
-            )
-
-        def op_fabs_overrider(arg):
-            return SymbolicExpr.create("fabs", SymbolicExpr.from_value(arg))
-
-        def op_ashr_overrider(lhs, rhs):
-            return SymbolicExpr.create(
-                "ashr",
-                SymbolicExpr.from_value(lhs),
-                SymbolicExpr.from_value(rhs),
-            )
-
-        def op_advance_overrider(ptr, offsets):
-            return SymbolicExpr.create(
-                "advance",
-                SymbolicExpr.from_value(ptr),
-                SymbolicExpr.from_value(offsets),
-            )
-
-        def op_fp_to_fp_overrider(src, dst_type, rounding_mode):
-            return SymbolicExpr.create(
-                "fp_to_fp",
-                SymbolicExpr.from_value(src),
-                dst_type,
-                rounding_mode,
-            )
-
-        def op_umulhi_overrider(lhs, rhs):
-            return SymbolicExpr.create(
-                "umulhi",
-                SymbolicExpr.from_value(lhs),
-                SymbolicExpr.from_value(rhs),
-            )
-
-        def op_cumsum_overrider(input, axis, reverse=False, dtype=None):
-            return SymbolicExpr.create(
-                "cumsum",
-                SymbolicExpr.from_value(input),
-                axis,
-                reverse,
-                dtype,
-            )
-
-        def op_bitcast_overrider(src, dst_type):
-            return SymbolicExpr.create(
-                "bitcast", SymbolicExpr.from_value(src), dst_type
-            )
-
-        def op_make_block_ptr_overrider(
-            base, shape, strides, offsets, tensor_shape, order
-        ):
-            raise NotImplementedError("MakeBlockPtr is not supported yet.")
-
-        def op_tensor_pointer_load_overrider(
-            ptr,
-            boundary_check,
-            padding_option,
-            cache_modifier,
-            eviction_policy,
-            is_volatile,
-        ):
-            raise NotImplementedError("TensorPointerLoad is not supported yet.")
-
-        def op_tensor_pointer_store_overrider(
-            ptr, value, boundary_check, cache_modifier, eviction_policy
-        ):
-            raise NotImplementedError("TensorPointerStore is not supported yet.")
-
-        # ── Concrete before-callbacks (active during concrete phase) ──
+        # Concrete before-callbacks (active during concrete phase)
 
         @self.lock_fn
         def before_load(ptr, mask, keys):
@@ -514,46 +309,6 @@ class RaceDetector(Client):
                 )
             )
 
-        # ── Map op types to callbacks ─────────────────────────────
-
-        OVERRIDER_MAP: dict[type[Op], Callable] = {
-            ProgramId: op_program_id_overrider,
-            RawLoad: op_raw_load_overrider,
-            Load: op_load_overrider,
-            RawStore: op_raw_store_overrider,
-            Store: op_store_overrider,
-            UnaryOp: op_unary_op_overrider,
-            BinaryOp: op_binary_op_overrider,
-            TernaryOp: op_ternary_op_overrider,
-            Dot: op_dot_overrider,
-            MakeRange: op_make_range_overrider,
-            AddPtr: op_addptr_overrider,
-            ExpandDims: op_expand_dims_overrider,
-            Broadcast: op_broadcast_overrider,
-            ReduceSum: op_reduce_sum_overrider,
-            ReduceMax: op_reduce_max_overrider,
-            ReduceMin: op_reduce_min_overrider,
-            Splat: op_splat_overrider,
-            MakeBlockPointer: op_make_block_ptr_overrider,
-            TensorPointerLoad: op_tensor_pointer_load_overrider,
-            TensorPointerStore: op_tensor_pointer_store_overrider,
-            Idiv: op_idiv_overrider,
-            Rsqrt: op_rsqrt_overrider,
-            CastImpl: op_cast_impl_overrider,
-            Reshape: op_reshape_overrider,
-            Trans: op_trans_overrider,
-            Join: op_join_overrider,
-            Fabs: op_fabs_overrider,
-            Ashr: op_ashr_overrider,
-            Advance: op_advance_overrider,
-            FpToFp: op_fp_to_fp_overrider,
-            Umulhi: op_umulhi_overrider,
-            CumSum: op_cumsum_overrider,
-            Bitcast: op_bitcast_overrider,
-            AtomicCas: op_atomic_cas_overrider,
-            AtomicRMW: op_atomic_rmw_overrider,
-        }
-
         BEFORE_MAP: dict[type[Op], Callable] = {
             Load: before_load,
             RawLoad: before_raw_load,
@@ -563,7 +318,7 @@ class RaceDetector(Client):
             AtomicCas: before_atomic_cas,
         }
 
-        overrider = OVERRIDER_MAP.get(op_type)
+        overrider = overrider_map.get(op_type)
         before = BEFORE_MAP.get(op_type)
 
         if overrider is not None:
@@ -576,130 +331,6 @@ class RaceDetector(Client):
             return callbacks
 
         return OpCallbacks()
-
-    def register_for_loop_callback(self) -> ForLoopCallbacks:
-        def _materialize_loop_value(expr: Any) -> int:
-            if isinstance(expr, SymbolicExpr):
-                if expr.op == "const":
-                    return SymbolicExprDataWrapper.coerce_int(expr.to_py())
-                elif expr.has_op("load"):
-                    self._need_concrete_fallback = True
-                    expr = expr.replace_subtree("load")
-                    return SymbolicExprDataWrapper.coerce_int(expr.to_py())
-                else:
-                    z3_expr, _ = expr.eval()
-                    if isinstance(z3_expr, IntNumRef):
-                        return z3_expr.as_long()
-                    self._need_concrete_fallback = True
-                    expr = expr.replace_subtree()
-                    return SymbolicExprDataWrapper.coerce_int(expr.to_py())
-            return int(expr)
-
-        @self.lock_fn
-        def _wrap_range(
-            iterable,
-            _lineno,
-            _range_type,
-            iter_args=None,
-            iter_kwargs=None,
-            _iter_callable=None,
-        ):
-            if self._phase != "symbolic":
-                return None
-            iter_args = tuple(iter_args or ())
-            iter_kwargs = iter_kwargs or {}
-            if isinstance(iterable, _RangeWrapper):
-                return iterable
-
-            args = tuple(SymbolicExpr.from_value(v) for v in iter_args)
-            if not args and iter_kwargs:
-                start_expr = SymbolicExpr.from_value(iter_kwargs.get("start", 0))
-                stop_expr = SymbolicExpr.from_value(
-                    iter_kwargs.get("stop", iter_kwargs.get("end"))
-                )
-                step_expr = SymbolicExpr.from_value(iter_kwargs.get("step", 1))
-                if stop_expr is not None:
-                    args = (start_expr, stop_expr, step_expr)
-
-            if not args and (
-                isinstance(iterable, range)
-                or (
-                    hasattr(iterable, "start")
-                    and hasattr(iterable, "stop")
-                    and hasattr(iterable, "step")
-                )
-            ):
-                args = tuple(
-                    SymbolicExpr.from_value(getattr(iterable, attr, default))
-                    for attr, default in (("start", 0), ("stop", 0), ("step", 1))
-                )
-
-            if not args:
-                return None
-
-            start_expr, stop_expr, step_expr = 0, None, 1
-            if len(args) == 1:
-                stop_expr = args[0]
-            elif len(args) == 2:
-                start_expr, stop_expr = args[0], args[1]
-            else:
-                start_expr, stop_expr, step_expr = args[0], args[1], args[2]
-
-            step = _materialize_loop_value(step_expr)
-            start = _materialize_loop_value(start_expr)
-            stop = _materialize_loop_value(stop_expr)
-            concrete_range = range(start, stop, step)
-            return _RangeWrapper(
-                concrete_range,
-                length=len(concrete_range),
-                start=start,
-                stop=stop,
-                step=step,
-            )
-
-        @self.lock_fn
-        def loop_hook_before(lineno, iterable):
-            if self._phase != "symbolic":
-                return
-            if not isinstance(iterable, _RangeWrapper):
-                return
-            idx_z3 = Int(f"loop_i_{lineno}")
-            sym = SymbolicExpr.create("const", idx_z3, tl.int32)
-            idx = tl.tensor(sym, tl.int32)
-            ctx = LoopContext(
-                lineno,
-                iterable.length,
-                idx,
-                idx_z3,
-                start=iterable.start,
-                stop=iterable.stop,
-                step=iterable.step,
-            )
-            sym.loop_ctx = ctx
-            self._loop_stack.append(ctx)
-
-        @self.lock_fn
-        def loop_hook_iter_overrider(lineno, idx):
-            if self._phase != "symbolic":
-                return idx
-            if self._loop_stack and self._loop_stack[-1].lineno == lineno:
-                return self._loop_stack[-1].idx
-            return idx
-
-        @self.lock_fn
-        def loop_hook_after(lineno: int) -> None:
-            if self._phase != "symbolic":
-                return
-            if not self._loop_stack or self._loop_stack[-1].lineno != lineno:
-                return
-            self._loop_stack.pop()
-
-        return ForLoopCallbacks(
-            range_wrapper_factory=_wrap_range,
-            before_loop_callback=loop_hook_before,
-            loop_iter_overrider=loop_hook_iter_overrider,
-            after_loop_callback=loop_hook_after,
-        )
 
     def finalize(self) -> list:
         if self._phase == "z3_done":
