+++ triton_viz/clients/sanitizer/sanitizer.py
@@ -1,4 +1,4 @@
-from collections.abc import Callable, Iterator
+from collections.abc import Callable
 from dataclasses import dataclass
 from functools import cached_property
 from typing import (
@@ -10,7 +10,6 @@ from typing import (
 )
 import sys
 
-import numpy as np
 from torch import Tensor
 from z3 import (
     Solver,
@@ -33,35 +32,9 @@ from ...core.data import (
     Load,
     RawStore,
     Store,
-    UnaryOp,
-    BinaryOp,
-    TernaryOp,
-    ProgramId,
-    Dot,
-    MakeRange,
-    AddPtr,
-    ExpandDims,
-    Broadcast,
-    ReduceSum,
-    ReduceMax,
-    ReduceMin,
-    Splat,
     MakeBlockPointer,
     TensorPointerLoad,
     TensorPointerStore,
-    Idiv,
-    Rsqrt,
-    CastImpl,
-    Reshape,
-    Trans,
-    Join,
-    Fabs,
-    Ashr,
-    Advance,
-    FpToFp,
-    Umulhi,
-    CumSum,
-    Bitcast,
     AtomicCas,
     AtomicRMW,
 )
@@ -72,9 +45,9 @@ from ..utils import (
 )
 from ..symbolic_engine import (
     SymbolicExpr,
-    SymbolicExprDataWrapper,
+    SymbolicClient,
+    RangeWrapper,
     PendingCheck,
-    LoopContext,
     Z3Expr,
     ConstraintConjunction,
 )
@@ -118,21 +91,6 @@ def _range_to_iterator_constraint(
     # `i in range(start, stop, step)` iff bounds hold and (i-start) is a multiple
     # of abs(step). Use a positive modulus to avoid negative-divisor semantics.
     return And(bounds, (var - start) % abs_step == 0)
-
-
-@dataclass
-class RangeWrapper:
-    iterable: Any
-    length: int
-    start: int
-    stop: int
-    step: int
-
-    def __iter__(self) -> Iterator[Any]:
-        return iter(self.iterable)
-
-    def __len__(self) -> int:
-        return self.length
 
 
 class Sanitizer(Client):
@@ -191,41 +149,6 @@ class Sanitizer(Client):
         raise NotImplementedError
 
 
-_UNARY_NUMPY_TO_SYM_OP: dict[Callable[..., Any], str] = {
-    np.cos: "cos",
-    np.exp: "exp",
-    np.exp2: "exp2",
-    np.abs: "abs",
-    np.floor: "floor",
-    np.ceil: "ceil",
-    np.log: "log",
-    np.log2: "log2",
-    np.sqrt: "sqrt",
-    np.sin: "sin",
-}
-
-_BINARY_NUMPY_TO_SYM_OP: dict[Callable[..., Any], str] = {
-    np.add: "add",
-    np.subtract: "sub",
-    np.multiply: "mul",
-    np.divide: "div",
-    np.less: "less",
-    np.less_equal: "less_equal",
-    np.greater: "greater",
-    np.greater_equal: "greater_equal",
-    np.not_equal: "not_equal",
-    np.equal: "equal",
-    np.fmod: "mod",
-    np.maximum: "maximum",
-    np.minimum: "minimum",
-    np.bitwise_and: "bitwise_and",
-    np.bitwise_or: "bitwise_or",
-    np.bitwise_xor: "bitwise_xor",
-    np.right_shift: "right_shift",
-    np.left_shift: "left_shift",
-}
-
-
 def _make_signature(
     addr_expr: Z3Expr,
     constraints: ConstraintConjunction,
@@ -266,9 +189,9 @@ class _FnSymbolicCache:
 _fn_symbolic_cache_set: set[_FnSymbolicCache] = set()
 
 
-class SymbolicSanitizer(Sanitizer):
+class SymbolicSanitizer(Sanitizer, SymbolicClient):
     def __init__(self, abort_on_error: bool = True):
-        super().__init__(abort_on_error=abort_on_error)  # Initialize parent class
+        super().__init__(abort_on_error=abort_on_error)
         self.records: list[OutOfBoundsRecordZ3] = []
         self.grid: Optional[tuple[int, ...]] = None
         self.grid_idx: Optional[tuple[int, ...]] = None
@@ -276,7 +199,6 @@ class SymbolicSanitizer(Sanitizer):
         self.tensor_addrs: list[tuple[int, int, Tensor]] = []
         self.tensor_names: dict[int, set[str]] = {}
         self.need_full_grid: Optional[bool] = None
-        self.loop_stack: list[LoopContext] = []
         self.last_grid: Optional[tuple[int, int, int]] = None
         self.cache_args: list[Any] = []
         self.cache_grid: Optional[tuple[int, ...]] = None
@@ -284,9 +206,6 @@ class SymbolicSanitizer(Sanitizer):
         self.pid_ok: Optional[BoolRef] = None
         self.solver: Optional[Solver] = None
         self.addr_sym: Optional[ArithRef] = None
-        SymbolicExpr.set_loop_ctx_provider(
-            lambda *_args, **_kwargs: self.loop_stack[-1] if self.loop_stack else None
-        )
 
     def _collect_tensor_base(self, expr: SymbolicExpr) -> Optional[int]:
         def walk(node: SymbolicExpr) -> Optional[int]:
@@ -406,11 +325,11 @@ class SymbolicSanitizer(Sanitizer):
         """
         # check memory access using z3
         z3_addr, z3_constraints = expr.eval()
-        if not self.loop_stack:
+        if not self._loop_stack:
             self._check_range_satisfiable(z3_addr, z3_constraints, expr)
             return
 
-        ctx = self.loop_stack[-1]
+        ctx = self._loop_stack[-1]
         signature = _make_signature(z3_addr, z3_constraints)
         pending_idx = ctx.signature_cache.get(signature)
         if pending_idx is None:
@@ -548,490 +467,201 @@ class SymbolicSanitizer(Sanitizer):
     def grid_idx_callback(self, grid_idx: tuple[int, ...]) -> None:
         self.grid_idx = grid_idx
 
+    # ── Sanitizer-specific operation overriders ─────────────────
+
+    def _op_raw_load_overrider(self, ptr, cache_modifier, eviction_policy, is_volatile):
+        return self._op_load_overrider(
+            ptr, None, None, cache_modifier, eviction_policy, is_volatile
+        )
+
+    def _op_load_overrider(self, ptr, mask, other, *args):
+        if isinstance(ptr, SymbolicExpr) and ptr.has_op("load"):
+            self.need_full_grid = True
+            ptr = ptr.replace_subtree("load")
+        if isinstance(mask, SymbolicExpr) and mask.has_op("load"):
+            self.need_full_grid = True
+            mask = mask.replace_subtree("load")
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        mask_sym = SymbolicExpr.from_value(mask) if mask is not None else None
+        other_sym = SymbolicExpr.from_value(other) if other is not None else None
+        ret = SymbolicExpr.create("load", ptr_sym, mask_sym, other_sym)
+        self._handle_access_check(ret)
+        return ret
+
+    def _op_raw_store_overrider(self, ptr, value, cache_modifier, eviction_policy):
+        return self._op_store_overrider(
+            ptr, value, None, cache_modifier, eviction_policy
+        )
+
+    def _op_store_overrider(self, ptr, value, mask, *args):
+        if isinstance(ptr, SymbolicExpr) and ptr.has_op("load"):
+            self.need_full_grid = True
+            ptr = ptr.replace_subtree("load")
+        if isinstance(mask, SymbolicExpr) and mask.has_op("load"):
+            self.need_full_grid = True
+            mask = mask.replace_subtree("load")
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        value_sym = SymbolicExpr.from_value(value)
+        mask_sym = SymbolicExpr.from_value(mask) if mask is not None else None
+        ret = SymbolicExpr.create("store", ptr_sym, value_sym, mask_sym)
+        self._handle_access_check(ret)
+        return ret
+
+    def _op_make_block_ptr_overrider(
+        self, base, shape, strides, offsets, tensor_shape, order
+    ):
+        base_sym = SymbolicExpr.from_value(base)
+        shape_syms = [SymbolicExpr.from_value(s) for s in shape]
+        stride_syms = [SymbolicExpr.from_value(s) for s in strides]
+        offset_syms = [SymbolicExpr.from_value(o) for o in offsets]
+        block_shape_vals = [int(b) for b in tensor_shape]
+        order_vals = [int(o) for o in order]
+        return SymbolicExpr.create(
+            "make_block_ptr",
+            base_sym,
+            shape_syms,
+            stride_syms,
+            offset_syms,
+            block_shape_vals,
+            order_vals,
+        )
+
+    def _op_tensor_pointer_load_overrider(
+        self,
+        ptr,
+        boundary_check,
+        padding_option,
+        cache_modifier,
+        eviction_policy,
+        is_volatile,
+    ):
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        ret = SymbolicExpr.create("tensor_pointer_load", ptr_sym, boundary_check)
+        self._handle_access_check(ret)
+        return ret
+
+    def _op_tensor_pointer_store_overrider(
+        self, ptr, value, boundary_check, cache_modifier, eviction_policy
+    ):
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        value_sym = SymbolicExpr.from_value(value)
+        ret = SymbolicExpr.create(
+            "tensor_pointer_store", ptr_sym, value_sym, boundary_check
+        )
+        self._handle_access_check(ret)
+        return ret
+
+    def _op_atomic_cas_overrider(self, ptr, cmp, val, sem, scope):
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        cmp_sym = SymbolicExpr.from_value(cmp)
+        val_sym = SymbolicExpr.from_value(val)
+        return SymbolicExpr.create("atomic_cas", ptr_sym, cmp_sym, val_sym)
+
+    def _op_atomic_rmw_overrider(self, rmwOp, ptr, val, mask, sem, scope):
+        ptr_sym = SymbolicExpr.from_value(ptr)
+        val_sym = SymbolicExpr.from_value(val)
+        mask_sym = SymbolicExpr.from_value(mask)
+        return SymbolicExpr.create("atomic_rmw", ptr_sym, val_sym, mask_sym)
+
     def register_op_callback(self, op_type: type[Op]) -> OpCallbacks:
-        def op_program_id_overrider(axis):
-            return SymbolicExpr.create("pid", axis)
+        overrider_map = self._build_op_overrider_map()
+        overrider_map.update(
+            {
+                RawLoad: self._op_raw_load_overrider,
+                Load: self._op_load_overrider,
+                RawStore: self._op_raw_store_overrider,
+                Store: self._op_store_overrider,
+                MakeBlockPointer: self._op_make_block_ptr_overrider,
+                TensorPointerLoad: self._op_tensor_pointer_load_overrider,
+                TensorPointerStore: self._op_tensor_pointer_store_overrider,
+                AtomicCas: self._op_atomic_cas_overrider,
+                AtomicRMW: self._op_atomic_rmw_overrider,
+            }
+        )
+        overrider = overrider_map.get(op_type)
+        if overrider is not None:
+            return OpCallbacks(op_overrider=self.lock_fn(overrider))
+        return OpCallbacks()
 
-        def op_raw_load_overrider(ptr, cache_modifier, eviction_policy, is_volatile):
-            return op_load_overrider(
-                ptr, None, None, cache_modifier, eviction_policy, is_volatile
-            )
+    # ── For-loop hook overrides ──────────────────────────────────
 
-        def op_load_overrider(ptr, mask, other, *args):
-            # deal with indirect loads
-            if isinstance(ptr, SymbolicExpr) and ptr.has_op("load"):
-                self.need_full_grid = True
-                ptr = ptr.replace_subtree("load")
+    def _on_data_dependent_value(self) -> None:
+        self.need_full_grid = True
 
-            if isinstance(mask, SymbolicExpr) and mask.has_op("load"):
-                self.need_full_grid = True
-                mask = mask.replace_subtree("load")
-
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            mask_sym = SymbolicExpr.from_value(mask) if mask is not None else None
-            other_sym = SymbolicExpr.from_value(other) if other is not None else None
-
-            ret = SymbolicExpr.create(
-                "load",
-                ptr_sym,
-                mask_sym,
-                other_sym,
-            )
-
-            # check memory access using z3 (defer in loops or check immediately)
-            self._handle_access_check(ret)
-            return ret
-
-        def op_raw_store_overrider(ptr, value, cache_modifier, eviction_policy):
-            return op_store_overrider(ptr, value, None, cache_modifier, eviction_policy)
-
-        def op_store_overrider(ptr, value, mask, *args):
-            # deal with indirect loads
-            if isinstance(ptr, SymbolicExpr) and ptr.has_op("load"):
-                self.need_full_grid = True
-                ptr = ptr.replace_subtree("load")
-
-            if isinstance(mask, SymbolicExpr) and mask.has_op("load"):
-                self.need_full_grid = True
-                mask = mask.replace_subtree("load")
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            value_sym = SymbolicExpr.from_value(value)
-            mask_sym = SymbolicExpr.from_value(mask) if mask is not None else None
-            ret = SymbolicExpr.create(
-                "store",
-                ptr_sym,
-                value_sym,
-                mask_sym,
-            )
-
-            # check memory access using z3 (defer in loops or check immediately)
-            self._handle_access_check(ret)
-            return ret
-
-        def op_unary_op_overrider(arg, op):
-            arg_sym = SymbolicExpr.from_value(arg)
-            try:
-                name = _UNARY_NUMPY_TO_SYM_OP[op]
-            except KeyError:
-                raise NotImplementedError(
-                    f"Unsupported unary operation: {op} on {arg_sym}"
-                )
-            return SymbolicExpr.create(name, arg_sym)
-
-        def op_binary_op_overrider(lhs, rhs, op):
-            lhs_sym = SymbolicExpr.from_value(lhs)
-            rhs_sym = SymbolicExpr.from_value(rhs)
-            try:
-                op_name = _BINARY_NUMPY_TO_SYM_OP[op]
-            except KeyError:
-                raise NotImplementedError(
-                    f"Unsupported binary operation: {op} between {lhs_sym} and {rhs_sym}"
-                )
-            return SymbolicExpr.create(op_name, lhs_sym, rhs_sym)
-
-        def op_ternary_op_overrider(lhs, rhs, other, op):
-            lhs_sym = SymbolicExpr.from_value(lhs)
-            rhs_sym = SymbolicExpr.from_value(rhs)
-            other = SymbolicExpr.from_value(other)
-            if op is np.where:
-                return SymbolicExpr.create("where", lhs_sym, rhs_sym, other)
-            else:
-                raise NotImplementedError(
-                    f"Unsupported ternary operation: {op} between {lhs_sym}, {rhs_sym} and {other}"
-                )
-
-        def op_addptr_overrider(ptr, offset):
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            offset_sym = SymbolicExpr.from_value(offset)
-            return SymbolicExpr.create("addptr", ptr_sym, offset_sym)
-
-        def op_dot_overrider(a, b, d, input_precision, max_num_imprecise_acc):
-            a_sym = SymbolicExpr.from_value(a)
-            b_sym = SymbolicExpr.from_value(b)
-            d_sym = SymbolicExpr.from_value(d) if d is not None else None
-            return SymbolicExpr.create("dot", a_sym, b_sym, d_sym)
-
-        def op_make_range_overrider(ret_ty, start, end):
-            return SymbolicExpr.create(
-                "arange",
-                SymbolicExpr.from_value(ret_ty),
-                SymbolicExpr.from_value(start),
-                SymbolicExpr.from_value(end),
-            )
-
-        def op_expand_dims_overrider(arg, axis):
-            return SymbolicExpr.create(
-                "expand_dims", SymbolicExpr.from_value(arg), axis
-            )
-
-        def op_broadcast_overrider(arg, shape):
-            return SymbolicExpr.create("broadcast", SymbolicExpr.from_value(arg), shape)
-
-        def op_reduce_sum_overrider(input, axis=None, keep_dims=False, **kwargs):
-            return SymbolicExpr.create(
-                "sum", SymbolicExpr.from_value(input), axis, keep_dims
-            )
-
-        def op_reduce_max_overrider(input, axis=None, keep_dims=False, **kwargs):
-            return SymbolicExpr.create(
-                "max", SymbolicExpr.from_value(input), axis, keep_dims
-            )
-
-        def op_reduce_min_overrider(input, axis=None, keep_dims=False, **kwargs):
-            return SymbolicExpr.create(
-                "min", SymbolicExpr.from_value(input), axis, keep_dims
-            )
-
-        def op_splat_overrider(shape, arg):
-            return SymbolicExpr.create("splat", shape, SymbolicExpr.from_value(arg))
-
-        def op_make_block_ptr_overrider(
-            base, shape, strides, offsets, tensor_shape, order
-        ):
-            base_sym = SymbolicExpr.from_value(base)
-            shape_syms = [SymbolicExpr.from_value(s) for s in shape]
-            stride_syms = [SymbolicExpr.from_value(s) for s in strides]
-            offset_syms = [SymbolicExpr.from_value(o) for o in offsets]
-            block_shape_vals = [int(b) for b in tensor_shape]
-            order_vals = [int(o) for o in order]
-            return SymbolicExpr.create(
-                "make_block_ptr",
-                base_sym,
-                shape_syms,
-                stride_syms,
-                offset_syms,
-                block_shape_vals,
-                order_vals,
-            )
-
-        def op_tensor_pointer_load_overrider(
-            ptr,
-            boundary_check,
-            padding_option,
-            cache_modifier,
-            eviction_policy,
-            is_volatile,
-        ):
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            ret = SymbolicExpr.create("tensor_pointer_load", ptr_sym, boundary_check)
-            self._handle_access_check(ret)
-            return ret
-
-        def op_tensor_pointer_store_overrider(
-            ptr, value, boundary_check, cache_modifier, eviction_policy
-        ):
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            value_sym = SymbolicExpr.from_value(value)
-            ret = SymbolicExpr.create(
-                "tensor_pointer_store", ptr_sym, value_sym, boundary_check
-            )
-            self._handle_access_check(ret)
-            return ret
-
-        def op_idiv_overrider(lhs, rhs):
-            result = SymbolicExpr.from_value(lhs) // SymbolicExpr.from_value(rhs)
-            return result
-
-        def op_rsqrt_overrider(arg):
-            return SymbolicExpr.create("rsqrt", SymbolicExpr.from_value(arg))
-
-        def op_cast_impl_overrider(src, dst_type):
-            return SymbolicExpr.create("cast_impl", src, dst_type)
-
-        def op_reshape_overrider(arg, shape, allow_reorder):
-            # For symbolic execution, we track the reshape operation
-            # arg is the input tensor, shape is the new shape, allow_reorder is a flag
-            arg_sym = SymbolicExpr.from_value(arg)
-            shape_sym = SymbolicExpr.from_value(shape)
-            # Create a reshape symbolic expression
-            return SymbolicExpr.create("reshape", arg_sym, shape_sym)
-
-        def op_join_overrider(lhs, rhs):
-            # Join operation combines two tensors along the last axis
-            lhs_sym = SymbolicExpr.from_value(lhs)
-            rhs_sym = SymbolicExpr.from_value(rhs)
-            # Create a join symbolic expression
-            return SymbolicExpr.create("join", lhs_sym, rhs_sym)
-
-        def op_fabs_overrider(arg):
-            arg_sym = SymbolicExpr.from_value(arg)
-            return SymbolicExpr.create("fabs", arg_sym)
-
-        def op_ashr_overrider(lhs, rhs):
-            lhs_sym = SymbolicExpr.from_value(lhs)
-            rhs_sym = SymbolicExpr.from_value(rhs)
-            return SymbolicExpr.create("ashr", lhs_sym, rhs_sym)
-
-        def op_advance_overrider(ptr, offsets):
-            # Advance operation for block pointers
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            offset_syms = [SymbolicExpr.from_value(o) for o in offsets]
-            return SymbolicExpr.create("advance", ptr_sym, offset_syms)
-
-        def op_umulhi_overrider(lhs, rhs):
-            lhs_sym = SymbolicExpr.from_value(lhs)
-            rhs_sym = SymbolicExpr.from_value(rhs)
-            return SymbolicExpr.create("umulhi", lhs_sym, rhs_sym)
-
-        def op_trans_overrider(arg, perm=[1, 0]):
-            return SymbolicExpr.create("trans", SymbolicExpr.from_value(arg), perm)
-
-        def op_cumsum_overrider(input, axis, reverse=False, dtype=None):
-            return SymbolicExpr.create(
-                "cumsum", SymbolicExpr.from_value(input), axis, reverse, dtype
-            )
-
-        def op_fp_to_fp_overrider(src, dst_type, rounding_mode):
-            return SymbolicExpr.create(
-                "fp_to_fp", SymbolicExpr.from_value(src), dst_type, rounding_mode
-            )
-
-        def op_bitcast_overrider(src, dst_type):
-            src_sym = SymbolicExpr.from_value(src)
-            return SymbolicExpr.create("bitcast", src_sym, dst_type)
-
-        def op_atomic_cas_overrider(ptr, cmp, val, sem, scope):
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            cmp_sym = SymbolicExpr.from_value(cmp)
-            val_sym = SymbolicExpr.from_value(val)
-            return SymbolicExpr.create("atomic_cas", ptr_sym, cmp_sym, val_sym)
-
-        def op_atomic_rmw_overrider(rmwOp, ptr, val, mask, sem, scope):
-            ptr_sym = SymbolicExpr.from_value(ptr)
-            val_sym = SymbolicExpr.from_value(val)
-            mask_sym = SymbolicExpr.from_value(mask)
-            return SymbolicExpr.create("atomic_rmw", ptr_sym, val_sym, mask_sym)
-
-        OP_TYPE_TO_OVERRIDER: dict[type[Op], Callable] = {
-            ProgramId: op_program_id_overrider,
-            RawLoad: op_raw_load_overrider,
-            Load: op_load_overrider,
-            RawStore: op_raw_store_overrider,
-            Store: op_store_overrider,
-            UnaryOp: op_unary_op_overrider,
-            BinaryOp: op_binary_op_overrider,
-            TernaryOp: op_ternary_op_overrider,
-            Dot: op_dot_overrider,
-            MakeRange: op_make_range_overrider,
-            AddPtr: op_addptr_overrider,
-            ExpandDims: op_expand_dims_overrider,
-            Broadcast: op_broadcast_overrider,
-            ReduceSum: op_reduce_sum_overrider,
-            ReduceMax: op_reduce_max_overrider,
-            ReduceMin: op_reduce_min_overrider,
-            Splat: op_splat_overrider,
-            MakeBlockPointer: op_make_block_ptr_overrider,
-            TensorPointerLoad: op_tensor_pointer_load_overrider,
-            TensorPointerStore: op_tensor_pointer_store_overrider,
-            Idiv: op_idiv_overrider,
-            Rsqrt: op_rsqrt_overrider,
-            CastImpl: op_cast_impl_overrider,
-            Reshape: op_reshape_overrider,
-            Trans: op_trans_overrider,
-            Join: op_join_overrider,
-            Fabs: op_fabs_overrider,
-            Ashr: op_ashr_overrider,
-            Advance: op_advance_overrider,
-            FpToFp: op_fp_to_fp_overrider,
-            Umulhi: op_umulhi_overrider,
-            CumSum: op_cumsum_overrider,
-            Bitcast: op_bitcast_overrider,
-            AtomicCas: op_atomic_cas_overrider,
-            AtomicRMW: op_atomic_rmw_overrider,
-        }
-
-        if op_type in OP_TYPE_TO_OVERRIDER:
-            return OpCallbacks(op_overrider=self.lock_fn(OP_TYPE_TO_OVERRIDER[op_type]))
-        else:
-            return OpCallbacks()
-
-    def register_for_loop_callback(self) -> ForLoopCallbacks:
-        def _materialize_loop_value(expr: Any) -> int:
-            if isinstance(expr, SymbolicExpr):
-                if expr.op == "const":
-                    return SymbolicExprDataWrapper.coerce_int(expr.to_py())
-                elif expr.has_op("load"):
-                    self.need_full_grid = True
-                    expr = expr.replace_subtree("load")
-                    return SymbolicExprDataWrapper.coerce_int(expr.to_py())
-                else:
-                    z3_expr, _ = expr.eval()
-                    if isinstance(z3_expr, IntNumRef):
-                        return z3_expr.as_long()
-                    self.need_full_grid = True
-                    expr = expr.replace_subtree()
-                    return SymbolicExprDataWrapper.coerce_int(expr.to_py())
-
-            return int(expr)
-
-        @self.lock_fn
-        def _wrap_range(
-            iterable,
-            _lineno,
-            _range_type,
-            iter_args=None,
-            iter_kwargs=None,
-            _iter_callable=None,
-        ):
-            """
-            Wrap range-like iterables so we can evaluate bounds once.
-            The caller can pass the original range call arguments to avoid
-            evaluating Python's built-in range on symbolic values.
-            """
-            iter_args = tuple(iter_args or ())
-            iter_kwargs = iter_kwargs or {}
-
-            if isinstance(iterable, RangeWrapper):
-                return iterable
-
-            args = tuple(SymbolicExpr.from_value(v) for v in iter_args)
-            # Prefer explicit args; fall back to kwargs when args are empty.
-            if not args and iter_kwargs:
-                start_expr = SymbolicExpr.from_value(iter_kwargs.get("start", 0))
-                stop_expr = SymbolicExpr.from_value(
-                    iter_kwargs.get("stop", iter_kwargs.get("end"))
-                )
-                step_expr = SymbolicExpr.from_value(iter_kwargs.get("step", 1))
-                if stop_expr is not None:
-                    args = (start_expr, stop_expr, step_expr)
-
-            if not args and (
-                isinstance(iterable, range)
-                or (
-                    hasattr(iterable, "start")
-                    and hasattr(iterable, "stop")
-                    and hasattr(iterable, "step")
-                )
-            ):
-                args = tuple(
-                    SymbolicExpr.from_value(getattr(iterable, attr, default))
-                    for attr, default in (("start", 0), ("stop", 0), ("step", 1))
-                )
-
-            if not args:
-                return None
-
-            start_expr, stop_expr, step_expr = 0, None, 1
-            if len(args) == 1:
-                stop_expr = args[0]
-            elif len(args) == 2:
-                start_expr, stop_expr = args[0], args[1]
-            else:
-                start_expr, stop_expr, step_expr = args[0], args[1], args[2]
-
-            step = _materialize_loop_value(step_expr)
-            start = _materialize_loop_value(start_expr)
-            stop = _materialize_loop_value(stop_expr)
-
-            concrete_range = range(start, stop, step)
-            length = len(concrete_range)
-
-            return RangeWrapper(
-                concrete_range, length=length, start=start, stop=stop, step=step
-            )
-
-        @self.lock_fn
-        def loop_hook_before(lineno, iterable):
-            if not isinstance(iterable, RangeWrapper):
-                if cfg.verbose:
-                    print(
-                        "not a range wrapper, skipping for-loop iterator association."
-                    )
-                return
-
-            idx_z3 = Int(f"loop_i_{lineno}")
-            sym = SymbolicExpr.create("const", idx_z3, tl.int32)
-            idx = tl.tensor(sym, tl.int32)
-            ctx = LoopContext(
-                lineno,
-                iterable.length,
-                idx,
-                idx_z3,
-                start=iterable.start,
-                stop=iterable.stop,
-                step=iterable.step,
-            )
-            sym.loop_ctx = ctx
-            self.loop_stack.append(ctx)
+    def _loop_hook_before(self, lineno, iterable):
+        if not isinstance(iterable, RangeWrapper):
             if cfg.verbose:
-                print(f"[Sanitizer] ▶ enter loop@{lineno}, len={iterable.length}")
+                print("not a range wrapper, skipping for-loop iterator association.")
+            return
+        super()._loop_hook_before(lineno, iterable)
+        if cfg.verbose:
+            print(f"[Sanitizer] ▶ enter loop@{lineno}, len={iterable.length}")
 
-        @self.lock_fn
-        def loop_hook_iter_overrider(lineno, idx):
-            if self.loop_stack and self.loop_stack[-1].lineno == lineno:
-                return self.loop_stack[-1].idx
-            return idx
+    def _loop_hook_after(self, lineno: int) -> None:
+        if not self._loop_stack or self._loop_stack[-1].lineno != lineno:
+            return
+        ctx = self._loop_stack.pop()
+        # Execute pending checks that were deferred during loop execution
+        solver = self.solver
+        addr_sym = self.addr_sym
+        assert solver is not None
+        assert addr_sym is not None
+        all_iterator_constraints: list[BoolRef] = []
+        if ctx.pending_checks:
+            solver.push()
+            # Add constraint for the current (innermost) loop
+            iterator_constraint = _range_to_iterator_constraint(
+                ctx.idx_z3, start=ctx.start, stop=ctx.stop, step=ctx.step
+            )
+            solver.add(iterator_constraint)
+            all_iterator_constraints.append(iterator_constraint)
 
-        @self.lock_fn
-        def loop_hook_after(lineno: int) -> None:
-            if not self.loop_stack or self.loop_stack[-1].lineno != lineno:
-                return
-            ctx = self.loop_stack.pop()
-            # Execute pending checks that were deferred during loop execution
-            solver = self.solver
-            addr_sym = self.addr_sym
-            assert solver is not None
-            assert addr_sym is not None
-            all_iterator_constraints: list[BoolRef] = []
-            if ctx.pending_checks:
-                solver.push()
-                # Add constraint for the current (innermost) loop
-                iterator_constraint = _range_to_iterator_constraint(
-                    ctx.idx_z3, start=ctx.start, stop=ctx.stop, step=ctx.step
+            # Also add constraints for all outer loops that are still active.
+            # This is critical for nested loops where the address expression
+            # depends on outer loop variables.
+            for outer_ctx in self._loop_stack:
+                outer_constraint = _range_to_iterator_constraint(
+                    outer_ctx.idx_z3,
+                    start=outer_ctx.start,
+                    stop=outer_ctx.stop,
+                    step=outer_ctx.step,
                 )
-                solver.add(iterator_constraint)
-                all_iterator_constraints.append(iterator_constraint)
+                solver.add(outer_constraint)
+                all_iterator_constraints.append(outer_constraint)
 
-                # Also add constraints for all outer loops that are still active.
-                # This is critical for nested loops where the address expression
-                # depends on outer loop variables.
-                for outer_ctx in self.loop_stack:
-                    outer_constraint = _range_to_iterator_constraint(
-                        outer_ctx.idx_z3,
-                        start=outer_ctx.start,
-                        stop=outer_ctx.stop,
-                        step=outer_ctx.step,
-                    )
-                    solver.add(outer_constraint)
-                    all_iterator_constraints.append(outer_constraint)
-
-            for pending_check in ctx.pending_checks:
-                addr_expr = pending_check.addr_expr
-                expr_constraints = pending_check.constraints
-                symbolic_expr = pending_check.symbolic_expr
-                # Use the source location captured when the check was created,
-                # not the current location (which would be the loop exit point)
-                source_location = pending_check.source_location
-
-                if cfg.verbose:
-                    print(
-                        "[Sanitizer] ▶ checking:",
-                        addr_expr,
-                        f" with iterator constraints: {all_iterator_constraints} ",
-                        f" and expression-related constraints: {expr_constraints} ",
-                    )
-
-                self._check_range_satisfiable(
-                    addr_expr,
-                    expr_constraints,
-                    symbolic_expr,
-                    source_location,
-                )
-            if ctx.pending_checks:
-                solver.pop()
+        for pending_check in ctx.pending_checks:
+            addr_expr = pending_check.addr_expr
+            expr_constraints = pending_check.constraints
+            symbolic_expr = pending_check.symbolic_expr
+            # Use the source location captured when the check was created,
+            # not the current location (which would be the loop exit point)
+            source_location = pending_check.source_location
 
             if cfg.verbose:
                 print(
-                    f"[Sanitizer] ▶ leave loop@{lineno} end. "
-                    f"(checked {len(ctx.pending_checks)} unique addr patterns)"
+                    "[Sanitizer] ▶ checking:",
+                    addr_expr,
+                    f" with iterator constraints: {all_iterator_constraints} ",
+                    f" and expression-related constraints: {expr_constraints} ",
                 )
 
-        return ForLoopCallbacks(
-            range_wrapper_factory=_wrap_range,
-            before_loop_callback=loop_hook_before,
-            loop_iter_overrider=loop_hook_iter_overrider,
-            after_loop_callback=loop_hook_after,
-        )
+            self._check_range_satisfiable(
+                addr_expr,
+                expr_constraints,
+                symbolic_expr,
+                source_location,
+            )
+        if ctx.pending_checks:
+            solver.pop()
+
+        if cfg.verbose:
+            print(
+                f"[Sanitizer] ▶ leave loop@{lineno} end. "
+                f"(checked {len(ctx.pending_checks)} unique addr patterns)"
+            )
+
+    def register_for_loop_callback(self) -> ForLoopCallbacks:
+        return SymbolicClient.register_for_loop_callback(self)
 
     def finalize(self) -> list:
         return []
